Initial commit
The script is currently in dry-run mode and exits after the first
watchlist iteration (50 pages), but should be ready for use by removing
the break (and maybe also removing the printing code). Usage:

    source venv/bin/activate
    ./unwatch [wiki [bot username [bot password]]]

All three arguments are optional and are read from standard input (in
this order) if omitted. Reading the password from stdin is recommended,
since process arguments are visible to everyone else on the system. The
wiki is a domain like en.wikipedia.org; the bot password must have view
and edit watchlist rights.